requirements.txt
gunicorn
pyyaml
seaborn
imageio
streamlit
imageio[ffmpeg]
xarray
cdsapi
regionmask
cartopy
netcdf4
scipy
boto3